+++ examples/expressions.js
@@ -85,7 +85,7 @@ console.log('\n2. USING FUNCTION MATH.PARSE');
 // parse an expression
 console.log('\nparse an expression into a node tree');
 var node1 = math.parse('sqrt(3^2 + 4^2)');
-print(node1.toString());                    // "ans = sqrt((3 ^ 2) + (4 ^ 2))"
+print(node1.toString());                    // "sqrt((3 ^ 2) + (4 ^ 2))"
 
 // compile the node
 var code1 = node1.compile(math);
@@ -97,7 +97,7 @@ print(code1.eval());                        // 5
 console.log('\nprovide a scope');
 var node2 = math.parse('x^a');
 var code2 = node2.compile(math);
-print(node2.toString());                    // "ans = x ^ a"
+print(node2.toString());                    // "x ^ a"
 var scope = {
   x: 3,
   a: 2