+++ HISTORY.md
@@ -8,14 +8,19 @@ https://github.com/josdejong/mathjs
 - Implemented function `unaryplus` and unary plus operator.
 - Added constant `phi`, the golden ratio (`phi = 1.618...`).
 - Added constant `version`, returning the version number of math.js as string.
+- Added unit `drop` (`gtt`).
 - Fixed not being able to load math.js using AMD/require.js.
 - Removed matrix support from conditional function `ifElse`.
 - Removed automatic assignment of expression results to variable `ans`. 
-  This functionality can be restored by preprocessing every evaluation, 
+  This functionality can be restored by pre- or postprocessing every evaluation, 
   something like:
   ```js
   function evalWithAns (expr, scope) {
-    return math.eval('ans=(' + expr + ')', scope);
+    var ans = math.eval(expr, scope);
+    if (scope) {
+      scope.ans = ans;
+    }
+    return ans;
   }
   ```